Lesson 27: Add Framebuffer and improve texture handling
Added a new Framebuffer class to improve graphics rendering, specifically for off-screen rendering.
Framebuffer class was added in the renderer package.
Modified the Texture class to handle generating a texture with a specific width and height without needing an existing file.
Modified SpriteRenderer and LevelEditorScene to refresh the texture of game objects after loading new files.
Also, added an equals method in the texture class to properly handle texture comparisons which improves visual consistency.
Improved texture indexing in RenderBatch.java for sprites using the new equals method and updated widow.java to create a new Framebuffer object and bind it to the render operation.